rename transformer to transformer.core
