+++ src/Transformer.Core/IEnvironmentProvider.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
-using Transformer.Model;
+using Transformer.Core.Model;
 
-namespace Transformer
+namespace Transformer.Core
 {
     public interface IEnvironmentProvider
     {
