+++ src/Transformer.Core/Logging/DebugLogFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using PowerDeploy.Common.Logging;
+using PowerDeploy.Transformer.Logging;
 
-namespace PowerDeploy.Transformer.Logging
+namespace Transformer.Core.Logging
 {
     /// <summary>
     /// Creates a Debug Logger, that logs all messages to: System.Diagnostics.Debug
