+++ src/Transformer.Core/Logging/ILog.cs
