+++ src/Transformer.Core/Logging/NullLogger.cs
@@ -1,33 +1,27 @@
 ﻿using System;
-using PowerDeploy.Common.Logging;
+using PowerDeploy.Transformer.Logging;
 
-namespace PowerDeploy.Transformer.Logging
+namespace Transformer.Core.Logging
 {
     /// <summary>
-    /// Default logger is to Console.WriteLine
+    /// Default logger is to System.Diagnostics.Debug.Print
     /// 
     /// Made public so its testable
     /// </summary>
     /// <remarks>https://github.com/ServiceStackV3/ServiceStackV3 BSD Licence.</remarks>
-    public class ConsoleLogger : ILog
+    public class NullLogger : ILog
     {
-        const string DEBUG = "DEBUG: ";
-        const string ERROR = "ERROR: ";
-        const string FATAL = "FATAL: ";
-        const string INFO = "INFO: ";
-        const string WARN = "WARN: ";
-
         /// <summary>
         /// Initializes a new instance of the <see cref="DebugLogger"/> class.
         /// </summary>
-        public ConsoleLogger(string type)
+        public NullLogger(string type)
         {
         }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DebugLogger"/> class.
         /// </summary>
-        public ConsoleLogger(Type type)
+        public NullLogger(Type type)
         {
         }
 
@@ -36,12 +30,6 @@ namespace PowerDeploy.Transformer.Logging
         /// </summary>
         private static void Log(object message, Exception exception)
         {
-            string msg = message == null ? string.Empty : message.ToString();
-            if (exception != null)
-            {
-                msg += ", Exception: " + exception.Message;
-            }
-            Console.WriteLine(msg);
         }
 
         /// <summary>
@@ -49,8 +37,6 @@ namespace PowerDeploy.Transformer.Logging
         /// </summary>
         private static void LogFormat(object message, params object[] args)
         {
-            string msg = message == null ? string.Empty : message.ToString();
-            Console.WriteLine(msg, args);
         }
 
         /// <summary>
@@ -58,93 +44,74 @@ namespace PowerDeploy.Transformer.Logging
         /// </summary>
         private static void Log(object message)
         {
-            string msg = message == null ? string.Empty : message.ToString();
-            Console.WriteLine(msg);
         }
 
         public void Debug(object message, Exception exception)
         {
-            Log(DEBUG + message, exception);
         }
 
         public void Debug(object message)
         {
-            Log(DEBUG + message);
         }
 
         public void DebugFormat(string format, params object[] args)
         {
-            LogFormat(DEBUG + format, args);
         }
 
         public void Print(object message)
         {
-            Log(message);
         }
 
         public void PrintFormat(string format, params object[] args)
         {
-            LogFormat(format, args);
         }
 
         public void Error(object message, Exception exception)
         {
-            Log(ERROR + message, exception);
         }
 
         public void Error(object message)
         {
-            Log(ERROR + message);
         }
 
         public void ErrorFormat(string format, params object[] args)
         {
-            LogFormat(ERROR + format, args);
         }
 
         public void Fatal(object message, Exception exception)
         {
-            Log(FATAL + message, exception);
         }
 
         public void Fatal(object message)
         {
-            Log(FATAL + message);
         }
 
         public void FatalFormat(string format, params object[] args)
         {
-            LogFormat(FATAL + format, args);
         }
 
         public void Info(object message, Exception exception)
         {
-            Log(INFO + message, exception);
         }
 
         public void Info(object message)
         {
-            Log(INFO + message);
         }
 
         public void InfoFormat(string format, params object[] args)
         {
-            LogFormat(INFO + format, args);
         }
 
         public void Warn(object message, Exception exception)
         {
-            Log(WARN + message, exception);
         }
 
         public void Warn(object message)
         {
-            Log(WARN + message);
         }
 
         public void WarnFormat(string format, params object[] args)
         {
-            LogFormat(WARN + format, args);
         }
     }
 }
