+++ src/Transformer.Core/Model/EnvironmentOverview.cs
@@ -1,6 +1,6 @@
 ﻿using System.Xml.Serialization;
 
-namespace Transformer.Model
+namespace Transformer.Core.Model
 {
     [XmlRoot("environment", Namespace = "")]
     public class EnvironmentOverview
