+++ src/Transformer.Core/Properties/AssemblyInfo.cs
