+++ src/Transformer.Core/StaticFolderEnvironmentLocator.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 
-namespace Transformer
+namespace Transformer.Core
 {
     public class StaticFolderEnvironmentLocator : IEnvironmentFolderLocator
     {
