+++ src/Transformer.Core/Template/VariableResolver.cs
@@ -3,13 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
-
-using PowerDeploy.Common.Logging;
 using PowerDeploy.Transformer.Logging;
-using Transformer.Logging;
-using Transformer.Model;
+using Transformer.Core.Logging;
+using Transformer.Core.Model;
 
-namespace PowerDeploy.Transformer.Template
+namespace Transformer.Core.Template
 {
     /// <summary>
     /// Variable Syntax:
