+++ src/Transformer.Core/packages.config
