+++ src/Transformer.PowerShell/ProtectEnvironments.cs
@@ -1,7 +1,8 @@
 ﻿using System.IO;
 using System.Management.Automation;
 using PowerDeploy.Transformer.Logging;
-using Transformer.Logging;
+using Transformer.Core;
+using Transformer.Core.Logging;
 
 namespace Transformer.PowerShell
 {
