+++ src/Transformer.Tests/EnvironmentSerializationTests.cs
@@ -5,8 +5,9 @@ using System.Xml;
 using System.Xml.Serialization;
 using NUnit.Framework;
 using PowerDeploy.Transformer;
-using Transformer.Model;
-using Environment = Transformer.Model.Environment;
+using Transformer.Core;
+using Transformer.Core.Model;
+using Environment = Transformer.Core.Model.Environment;
 
 namespace Transformer.Tests
 {
