+++ src/Transformer.Tests/TemplateEngineTests.cs
@@ -2,8 +2,9 @@
 using Moq;
 using NUnit.Framework;
 using PowerDeploy.Transformer;
-using PowerDeploy.Transformer.Template;
-using Transformer.Model;
+using Transformer.Core;
+using Transformer.Core.Model;
+using Transformer.Core.Template;
 
 namespace Transformer.Tests
 {
