+++ src/Transformer.Tests/Transformer.Tests.csproj
@@ -59,13 +59,13 @@
     <Compile Include="ZipPackagesTests.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Transformer\Transformer.csproj">
-      <Project>{b7a5ffd2-a16d-482c-8d28-4552b3d29aa0}</Project>
-      <Name>Transformer</Name>
-    </ProjectReference>
+    <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="packages.config" />
+    <ProjectReference Include="..\Transformer.Core\Transformer.Core.csproj">
+      <Project>{b7a5ffd2-a16d-482c-8d28-4552b3d29aa0}</Project>
+      <Name>Transformer.Core</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
