+++ src/Transformer.Tests/VariableResolverTests.cs
@@ -2,8 +2,8 @@
 using System.Linq;
 using NUnit.Framework;
 using PowerDeploy.Transformer;
-using PowerDeploy.Transformer.Template;
-using Transformer.Model;
+using Transformer.Core.Model;
+using Transformer.Core.Template;
 
 namespace Transformer.Tests
 {
