+++ src/Transformer.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.30723.0
+VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transformer.Tests", "Transformer.Tests\Transformer.Tests.csproj", "{654DEFFE-1B0B-418A-99E3-31C714A83627}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transformer", "Transformer\Transformer.csproj", "{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transformer.MsBuild.v4.0", "Transformer.MsBuild\Transformer.MsBuild.v4.0.csproj", "{7BC5D9DD-8D4A-455C-B4FA-4037157D7A6D}"
 EndProject
@@ -22,6 +20,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "nuget", "nuget", "{A72B4CFF
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transformer.PowerShell", "Transformer.PowerShell\Transformer.PowerShell.csproj", "{7FCADFB9-C69C-49A4-9A47-841BBF8950E7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transformer.Core", "Transformer.Core\Transformer.Core.csproj", "{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -35,12 +35,6 @@ Global
 		{654DEFFE-1B0B-418A-99E3-31C714A83627}.Package|Any CPU.Build.0 = Release|Any CPU
 		{654DEFFE-1B0B-418A-99E3-31C714A83627}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{654DEFFE-1B0B-418A-99E3-31C714A83627}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Package|Any CPU.ActiveCfg = Release|Any CPU
-		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Package|Any CPU.Build.0 = Release|Any CPU
-		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7BC5D9DD-8D4A-455C-B4FA-4037157D7A6D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7BC5D9DD-8D4A-455C-B4FA-4037157D7A6D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7BC5D9DD-8D4A-455C-B4FA-4037157D7A6D}.Package|Any CPU.ActiveCfg = Release|Any CPU
@@ -65,6 +59,12 @@ Global
 		{7FCADFB9-C69C-49A4-9A47-841BBF8950E7}.Package|Any CPU.Build.0 = Release|Any CPU
 		{7FCADFB9-C69C-49A4-9A47-841BBF8950E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7FCADFB9-C69C-49A4-9A47-841BBF8950E7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Package|Any CPU.ActiveCfg = Release|Any CPU
+		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Package|Any CPU.Build.0 = Release|Any CPU
+		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B7A5FFD2-A16D-482C-8D28-4552B3D29AA0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
